Move compilation utilities out of amdllpc.cpp. NFC.
-  Update documentation.
-  Don't use global variables outside of `amdllpc.cpp`.

Required for https://github.com/GPUOpen-Drivers/llpc/pull/1372.
